driver/requirements.txt
alabaster==0.7.16
babel==2.16.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
docutils==0.17.1
et-xmlfile==1.1.0
google-auth==2.34.0
idna==3.8
imagesize==1.4.1
Jinja2==3.1.4
kubernetes==25.3.0
MarkupSafe==2.1.5
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.3.2
pyaml==21.10.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
sdcclient==0.17.3
setuptools==80.9.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
TatSu==5.8.3
tzdata==2025.2
urllib3==2.2.2
websocket-client==1.8.0